fix hiệu năng web
