+++ fe/src/constants/universities.js
@@ -1,0 +1,25 @@
+export const UNIVERSITIES = [
+    "Đại học Bách khoa Hà Nội",
+    "Đại học Bách khoa TP.HCM",
+    "Đại học Cần Thơ",
+    "Đại học Công nghiệp TP.HCM",
+    "Đại học Đà Nẵng",
+    "Đại học Duy Tân",
+    "Đại học FPT",
+    "Đại học Huế",
+    "Đại học Kinh tế - Luật TP.HCM",
+    "Đại học Kinh tế (ĐHQGHN)",
+    "Đại học Kinh tế Quốc dân",
+    "Đại học Khoa học Tự nhiên TP.HCM",
+    "Đại học Khoa học Xã hội và Nhân văn TP.HCM",
+    "Đại học Lạc Hồng",
+    "Đại học Ngân hàng TP.HCM",
+    "Đại học Ngoại ngữ - Tin học TP.HCM",
+    "Đại học Ngoại thương",
+    "Đại học Nguyễn Tất Thành",
+    "Đại học Nông Lâm TP.HCM",
+    "Đại học Quốc gia Hà Nội",
+    "Đại học Quốc gia TP.HCM",
+    "Đại học Sư phạm TP.HCM",
+    "Đại học Tôn Đức Thắng"
+];
